Renamed Sector to MapRegion for possibility of different kinds of maps other than star maps.

+++ HexMap/src/HexMap.cpp
@@ -4,7 +4,7 @@
 
 namespace hexmap
 {
-    hex_set CreateFTRVMap(int width, int height)
+    hex_set CreateFTRVMap(size_t width, size_t height)
     {
         hex_set hexmap;
 
@@ -12,7 +12,7 @@ namespace hexmap
         {
             int q_offset = std::floor( (q+1)/2);
             
-            for(int r = -q_offset; r < height - q_offset; ++r)
+            for(int r = -q_offset; r < int(height) - q_offset; ++r)
             {
                 hexmap.insert(Hex(q, r, -q-r) );
             }
